refactor: move ingestion dummy traffic logic in own file
