+++ App/Logic/DataUploadLogic.swift
@@ -266,117 +266,6 @@ extension Logic.DataUpload {
   }
 }
 
-// MARK: - Dummy traffic
-
-extension Logic.DataUpload {
-  /// Updates the dummy ingestion traffic opportunity window taking the parameters from the Configuration and the RNGs.
-  struct UpdateDummyTrafficOpportunityWindowIfExpired: AppSideEffect {
-    func sideEffect(_ context: SideEffectContext<AppState, AppDependencies>) throws {
-      let state = context.getState()
-
-      let now = context.dependencies.now()
-
-      guard now > state.ingestion.dummyTrafficOpportunityWindow.windowEnd else {
-        // Window not started or finished yet.
-        return
-      }
-
-      try context.awaitDispatch(UpdateDummyTrafficOpportunityWindow())
-    }
-  }
-
-  /// Updates the dummy ingestion traffic opportunity window taking the parameters from the Configuration and the RNGs.
-  struct UpdateDummyTrafficOpportunityWindow: AppSideEffect {
-    func sideEffect(_ context: SideEffectContext<AppState, AppDependencies>) throws {
-      let state = context.getState()
-
-      let dummyTrafficStochasticDelay = context.dependencies.exponentialDistributionGenerator
-        .exponentialRandom(with: state.configuration.dummyIngestionMeanStochasticDelay)
-
-      try context.awaitDispatch(
-        SetDummyTrafficOpportunityWindow(
-          dummyTrafficStochasticDelay: dummyTrafficStochasticDelay,
-          dummyTrafficOpportunityWindowDuration: state.configuration.dummyIngestionWindowDuration,
-          now: context.dependencies.now()
-        )
-      )
-    }
-  }
-
-  /// Schedules a sequence of dummy ingestion requests for some random point in the future.
-  /// -seeAlso: https://github.com/immuni-app/immuni-documentation/blob/master/Traffic%20Analysis%20Mitigation.md
-  struct ScheduleDummyIngestionSequenceIfNecessary: AppSideEffect {
-    func sideEffect(_ context: SideEffectContext<AppState, AppDependencies>) throws {
-      let state = context.getState()
-
-      guard !state.ingestion.dummyTrafficSequenceScheduledInSession else {
-        // There is already another sequence schedule for this session.
-        return
-      }
-
-      // Reset the flag in case this is a new session
-      try context.awaitDispatch(SetDummyTrafficSequenceCancelled(value: false))
-
-      guard state.ingestion.dummyTrafficOpportunityWindow.contains(context.dependencies.now()) else {
-        // Not within the opportunity window. Nothing to do.
-        return
-      }
-
-      // Prevent other ingestion sequences from being scheduled in this session
-      try context.awaitDispatch(SetDummyIngestionSequenceScheduledForThisSession(value: true))
-
-      let startDelay = context.dependencies.exponentialDistributionGenerator
-        .exponentialRandom(with: state.configuration.dummyIngestionAverageStartUpDelay)
-
-      // Dispatch a delayed action and return
-      context.dispatch(StartIngestionSequenceIfNotCancelled().deferred(of: startDelay))
-    }
-  }
-
-  /// Attempts a simulated sequence of ingestion requests, unless it was cancelled for outside in the meanwhile.
-  /// -seeAlso: https://github.com/immuni-app/immuni-documentation/blob/master/Traffic%20Analysis%20Mitigation.md
-  struct StartIngestionSequenceIfNotCancelled: AppSideEffect {
-    func sideEffect(_ context: SideEffectContext<AppState, AppDependencies>) throws {
-      var executions = 0
-      while true {
-        // Fetch the state at every run because it might have been updated in the meanwhile
-        let state = context.getState()
-
-        guard !state.ingestion.isDummyTrafficSequenceCancelled else {
-          // The user signaled the intent of starting a genuine upload.
-          break
-        }
-
-        #warning("Define proper dummy request")
-        try? await(context.dependencies.networkManager.uploadData(body: .dummy(), otp: .init()))
-
-        let diceRoll = context.dependencies.uniformDistributionGenerator.randomNumberBetweenZeroAndOne()
-        let threshold = state.configuration.dummyIngestionRequestProbabilities[safe: executions]
-          ?? state.configuration.dummyIngestionRequestProbabilities.last
-          ?? AppLogger.fatalError("No probabilities defined")
-
-        guard diceRoll < threshold else {
-          // The generated probability was above the threshold.
-          break
-        }
-
-        let interRequestDelay = context.dependencies.exponentialDistributionGenerator
-          .exponentialRandom(with: state.configuration.dummyIngestionAverageRequestWaitingTime)
-
-        // Await for a random time
-        try await(Promise<Void>.deferring(of: interRequestDelay))
-        executions += 1
-      }
-
-      // Allow the scheduling of another session.
-      try context.awaitDispatch(SetDummyIngestionSequenceScheduledForThisSession(value: false))
-
-      // Update the opportunity window
-      try context.awaitDispatch(UpdateDummyTrafficOpportunityWindow())
-    }
-  }
-}
-
 // MARK: - StateUpdaters
 
 extension Logic.DataUpload {
@@ -402,45 +291,6 @@ extension Logic.DataUpload {
     func updateState(_ state: inout AppState) {
       state.ingestion.lastOtpValidationFailedAttempt = nil
       state.ingestion.otpValidationFailedAttempts = 0
-    }
-  }
-
-  /// Sets the opportunity window for the dummy ingestion traffic
-  struct SetDummyTrafficOpportunityWindow: AppStateUpdater {
-    let dummyTrafficStochasticDelay: Double
-    let dummyTrafficOpportunityWindowDuration: Double
-    let now: Date
-
-    func updateState(_ state: inout AppState) {
-      let windowStart = self.now.addingTimeInterval(self.dummyTrafficStochasticDelay)
-      let windowDuration = self.dummyTrafficOpportunityWindowDuration
-      state.ingestion.dummyTrafficOpportunityWindow = .init(windowStart: windowStart, windowDuration: windowDuration)
-    }
-  }
-
-  /// Marks a dummy traffic sequence as cancelled
-  struct SetDummyTrafficSequenceCancelled: AppStateUpdater {
-    let value: Bool
-
-    func updateState(_ state: inout AppState) {
-      state.ingestion.isDummyTrafficSequenceCancelled = self.value
-    }
-  }
-
-  /// Signals that a dummy traffic sequence is already scheduled for this session
-  struct SetDummyIngestionSequenceScheduledForThisSession: AppStateUpdater {
-    let value: Bool
-
-    func updateState(_ state: inout AppState) {
-      state.ingestion.dummyTrafficSequenceScheduledInSession = self.value
-    }
-  }
-
-  /// Handles the end of a foreground session
-  struct MarkForegroundSessionFinished: AppStateUpdater {
-    func updateState(_ state: inout AppState) {
-      state.ingestion.dummyTrafficSequenceScheduledInSession = false
-      state.ingestion.isDummyTrafficSequenceCancelled = true
     }
   }
 }
